Refactor asa tests for better isolation (#690)
* rev docs, module strings; refactor tests for better isolation (skip if extras not installed)
