+++ compiler/quilt/test/utils.py
@@ -38,6 +38,16 @@ def quilt_dev_mode(func):
             quilt._DEV_MODE = dev_mode
     return decorated
 
+def try_require(string):
+    """return true iff specified require string resolves properly;
+    for use with @pytest.mark.skipif"""
+    import pkg_resources
+    try:
+        pkg_resources.require(string)
+    except pkg_resources.ResolutionError as ex:
+        print(ex)
+        return False
+    return True
 
 class BasicQuiltTestCase(unittest.TestCase):
     """
